update makefile to build at debug/release directory; fixed some compile error

+++ insert_sort.cpp
@@ -9,9 +9,9 @@
 void insert_sort(vector<uint32_t> &vecInt)
 {
     // repeat (n - 1) times
-    for (size_t i = 1; i < vecInt.size(); ++i){
+    for (int i = 1; i < (int)vecInt.size(); ++i){
         uint32_t nKey = vecInt[i];
-        size_t j = i - 1;
+        int j = i - 1;
         // look back for the insert position
         while ((j >= 0) && (vecInt[j] > nKey)){
             vecInt[j + 1] = vecInt[j];
